requirements.txt
asgiref==3.7.2
cffi==1.15.1
cryptography==41.0.3
Django==4.2.4
django-rest-framework==0.1.0
djangorestframework==3.14.0
pycparser==2.21
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
